feat: 자동차들(Cars)에 레이싱 시도를 하는 기능 구현
- 기존에 테스트 완료 된 랜덤수 생성 기능과 입력 수에 따라 전진/정지 하는 기능을 합쳐서 레이싱 기능 구현
- RacingGameController 에 있던 랜덤수 생성 기능을 Cars 클래스 내부로 이동함
    위 기능 메서드 이동에 따른 테스트케이스 이동
