+++ src/main/java/controller/RacingGameController.java
@@ -22,11 +22,6 @@ public class RacingGameController {
         }
     }
 
-    public int makeRandomIntRangeZeroToNine() {
-        Random random = new Random();
-        return random.nextInt(9);
-    }
-
     public void initCars() {
         try {
             System.out.println("경주할 자동차 이름을 입력하세요. (이름은 쉼포(,) 기준으로 구분)");
