+++ src/test/java/model/CarTest.java
@@ -77,4 +77,15 @@ public class CarTest {
         List<String> champNames = cars.getChampions();
         assertThat(champNames).containsExactly("car1", "car3");
     }
+
+    @DisplayName("0~ 9 사이의 랜덤수를 만든다.")
+    @Test
+    public void 전진조건을_결정하는_랜덤수_유효성_Test() {
+        int i = 0;
+        while (i++ < 50) {
+            int number = cars.makeRandomIntRangeZeroToNine();
+            assertThat(number).isGreaterThanOrEqualTo(0);
+            assertThat(number).isLessThanOrEqualTo(9);
+        }
+    }
 }
